feat :  Business can be now bought using the company market app in phone (money gets subtracted from bank).     Business App now is fully functional
    Made int cash to long long cash and Bank_money to long long.

+++ include/phone/Bank_App.hpp
@@ -13,7 +13,7 @@ public:
 	void draw(sf::RenderWindow& window, const sf::Font& font);
 
 	bool is_Bank_App_Open = false;
-	static int Bank_Money;
+	static long long Bank_Money;
 
 
 private:
